updating the vector stuff

+++ entity-component-system/Vector3.h
@@ -3,12 +3,12 @@
 
 struct vec3
 {
-    double m_x;
-    double m_y;
-    double m_z;
+    float m_x;
+    float m_y;
+    float m_z;
 
     vec3() : m_x(0), m_y(0), m_z(0) { }
-    vec3(double x, double y, double z) : m_x(x), m_y(y), m_z(z) { }
+    vec3(float x, float y, float z) : m_x(x), m_y(y), m_z(z) { }
 };
 
 #endif
